Change minDigits and maxDigits + Refactoring
