+++ src/components/AbstractCounterDigit.js
