+++ src/components/AnimatedCounterDigit.js
