+++ src/components/CounterSegment.js
