+++ src/components/CounterSegmentSeparator.js
