+++ src/components/StaticCounterDigit.js
