+++ src/helpers/globalIntervals.js
