+++ test/unit/.eslintrc.json
@@ -1,5 +1,6 @@
 {
   "env": {
+    "jasmine": true,
     "jest": true
   }
 }
